pyproject.toml
[tool.poetry]
name = "musichat"
version = "0.1.0"
description = ""
authors = ["shutov.da <shutov.da@infotek.tu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.110.0"}
uvicorn = "^0.28.0"
pydantic-settings = "^2.2.1"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
sqlalchemy-utils = "^0.41.2"
uvloop = "^0.19.0"
sqlalchemy = {version = "^2.0.29", extras = ["asyncio"]}
phonenumbers = "^8.13.33"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
